requirements.txt
PySide6==6.5.3
pypdf==3.16.1
python-docx==0.8.11
pyinstaller>=6.0.0
